requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2
django-filter==21.1
django-import-export==2.7.1
djangorestframework==3.8.0
et-xmlfile==1.1.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.22.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
